We have a teapot! Wireframe and non-wireframe
git-svn-id: svn+ssh://njoubert.com/var/svn@200 997b1fe7-5dcb-cc6a-e049-b7da040a1fbe

+++ CS184_as5/trunk/NielsDebug.cpp
@@ -11,7 +11,7 @@
 
 using namespace std;
 
-#define printDebug(A, B)	if (NIELSDEBUG >= A) { cout << "  "; cout << clock()*1000/CLOCKS_PER_SEC; cout <<"ms: " << __FILE__ << ":" << __LINE__ << ":" << __FUNCTION__ << ": " << B << endl; }
+#define printDebug(A, B)	if (NIELSDEBUG >= A) { cout << "  "; cout << ((long)clock())*((double)1000/CLOCKS_PER_SEC); cout <<"ms: " << __FILE__ << ":" << __LINE__ << ":" << __FUNCTION__ << ": " << B << endl; }
 #define printStartDebug(A)	if (NIELSDEBUG) { cout << __FILE__ << ":" << __LINE__ << ":" << __FUNCTION__ << ": " << A; }
 #define printPartialDebug(A)	if (NIELSDEBUG) { cout << A; }
